Se actualiza proyecto con la informacion dada por el instructor con el CRUD completo ya la aplicacion de Rol y Dependencia

+++ src/main/java/com/project/dao/IUsuarioDao.java
@@ -3,6 +3,6 @@ package com.project.dao;
 import org.springframework.data.repository.CrudRepository;
 import com.project.entities.Usuario;
 
-public interface IUsuarioDao extends CrudRepository<Usuario, Long> {
+public interface IUsuarioDao extends CrudRepository<Usuario, Integer> {
 
 }
